pyproject.toml
[project]
name = "philosopher-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.119.1",
    "markitdown[docx]>=0.1.3",
    "matplotlib>=3.10.6",
    "networkx>=3.5",
    "nuitka>=2.8.9",
    "openai>=2.7.1",
    "pydantic>=2.12.3",
    "pytest>=9.0.0",
    "python-dotenv>=1.2.1",
    "fastmcp>=2.14.1",
    "mcp>=1.25.0",
    "ruff>=0.14.11",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
view = [
    "pyvis>=0.3.2",
    "streamlit>=1.52.2",
]
